Seat and auditorium rollback when creation fails #11
When the creation of a seat fails the seats already created and the auditorium are deleted to ensure that when an auditorium is created all seats are also created.
A test is added to verify this behavior. In this tests firsts a seat is created and then a seat fails. 
Some fixes in the other states are also carried on involving the creation of seats in tests.