Add demo pages
Add page for dashboard demo and for reserving tickets demo.
The former receives the dashboard data when opens and when a is ticket create or update.
The latter receives the full information of all tickets of the corresponding presentation when open and when a new ticket is created for that presentation